qrtr: Initialize sockaddr_qrtr for node 1
As node 0 is made a valid id in the kernel we're not allowed to specify
node 0 when binding our sockets. For now just hard code 1, as that's
where we are.

Either the kernel should define a reserved node id or we should have a
convenient way of specifying the local node, but for now this is good
enough.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>