ns: Implement DEL_CLIENT
Upon receiving a DEL_CLIENT some local or remote client has been
disconnected. If this is a local server we remove the server from the
lists and announce the removal to the remotes. Otherwise we need to
inform local servers that one of their clients might be gone, so we
propagate a DEL_SERVER message to all locally registered services.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>