libqrtr: Don't require socket to be bound for lookup
There's no reason to require that the socket to be bound to a specific
port in order to register a lookup, the transmission of the lookup
request will autobind the socket to a port for us.

Also drop the node and port from the payload, as qrtr-ns will register
the lookup on the source sq, rather than the passed port.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>