qrtr-ns: Move DEL_SERVER logic to server_del
When a server is disappearing, either from a DEL_SERVER, BYE or
DEL_CLIENT message the same messages needs to be sent; to local
observers and through broadcasts to remote nodes.

So move the logic from the DEL_SERVER command handler to server_del, in
order to invoke this regardless of reason for the server's
disappearance.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>