libqrtr: Don't enforce server port to be bound
When sending a new_server message to the qrtr-ns the node and port in
the packet is ignored (to not allow clients spoofing the information),
so there's no reason to have this artificial check in the
qrtr_new_server() implementation either.

The socket will be bound to a port automatically when the new_server
message is sent out.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>