libqrtr: Correct "instance" composition
The "instance" id of a service is a composition of the instance id and
the version. In some plaeces in the library I made this two half words
instead of the actually expected 8 bits version + 24 bit instance. Fix
this in the library.

This was not spotted previously because all prior work has dealt with
instance id 0, but with the QRTR support in Diag this is now an issue.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>