qrtr-lookup: print better information about DIAG service
The DIAG service has its own encoding for the instance number which
is unrelated to the version. This means that the information we
show to the user in the version and instance columns of the table is
parsed incorrectly and often useless. Instead, treat the instance
number as special if the service is a DIAG service and print out a
more meaningful interpretation.

Signed-off-by: Eric Caruso <ejcaruso@chromium.org>