lib: meson.build: restore libqrtr SONAME
The previous version of `libqrtr` had `libqrtr.so.1` as their SONAME,
while the current version have only `libqrtr.so`. This is problematic as
previously built binaries using this lib will have to be rebuilt for
this new version (or rather, re-linked, which is practically
equivalent).

This change ensures the SONAME is kept by setting the project version
in the top-level `meson.build` and uses this string as the shared
library version.

In practice, the generated library filename is now `libqrtr.so.1.1` with
SONAME still being `libqrtr.so.1`.