meson: Specify project and library version
Specify project version in meson project() call.
Withouth this in generated qrtr.pc pkg-config file
version will be undefined, and this is a packaging error.
E.g. Alpine Linux's build system doesn't allow this to pass.

Specify also shared library soversion.
This reults in library being called libqrtr.so.1.1 and
symlinks generated and installed as:

 libqrtr.so -> libqrtr.so.1
 libqrtr.so.1 -> libqrtr.so.1.1

This is important for -dev subpackage.

Signed-off-by: Alexey Minnekhanov <alexey.min@gmail.com>